관리자페이지 요양사, 게시글 댓글부분  하는중

+++ silverShelter/src/main/java/com/silver/shelter/board/model/mapper/BoardMapper.java
@@ -131,9 +131,10 @@ public interface BoardMapper {
 	/** 댓글 전체 조회
 	 * @return
 	 */
-	List<Comment> commentAllSelect(int boardNo);
+	List<Comment> commentAllSelect(long boardNo);
 
 
+
